Meta: Move BuildVcpkg script to the Meta directory
It was the only thing left in the Toolchain directory.
